# Conflicts:
#	controllers/tweetController.js

+++ controllers/tweetController.js
@@ -23,13 +23,18 @@ const tweetController = {
         userAvatar: tweet.User.avatar,
         replyLength: tweet.Replies.length,
         likeLength: tweet.Likes.length,
-        isLiked: req.user.LikedTweets.map(likeTweet => likeTweet.id).includes(
-          tweet.id
-        )
+        isLiked: helpers
+          .getUser(req)
+          .LikedTweets.map(likeTweet => likeTweet.id)
+          .includes(tweet.id)
       }))
-      return res.render('tweets', { reorganizationTweets, popularUser })
+      return res.render('tweets', {status: (200), reorganizationTweets, popularUser })
     } catch (err) {
-      console.warn(err)
+      console.log(err)
+      res.status(302);
+      console.log('getTweets err')
+      req.flash('error_messages', '讀取貼文串失敗！')
+      return res.redirect('back')
     }
   },
   addTweet: async (req, res) => {
@@ -45,11 +50,16 @@ const tweetController = {
       }
       const tweet = await Tweet.create({
         description,
-        UserId: req.user.id
+        UserId: helpers.getUser(req).id
       })
+      res.status(200)
       return res.redirect('/tweets')
     } catch (err) {
-      console.warn(err)
+      console.log(err)
+      res.status(302);
+      console.log('addTweet err')
+      req.flash('error_messages', '新增留言失敗！')
+      return res.redirect('back')
     }
   },
   getTweet: async (req, res) => {
@@ -67,16 +77,21 @@ const tweetController = {
       tweetJson.year = dayjs(`${tweetJson.createdAt}`).format('YYYY')
       tweetJson.month = dayjs(`${tweetJson.createdAt}`).format('M')
       tweetJson.day = dayjs(`${tweetJson.createdAt}`).format('D')
-      tweetJson.isLiked = req.user.LikedTweets.map(
-        likeTweet => likeTweet.id
-      ).includes(tweetJson.id)
+      tweetJson.isLiked = helpers
+        .getUser(req)
+        .LikedTweets.map(likeTweet => likeTweet.id)
+        .includes(tweetJson.id)
       const tweetReplies = tweetJson.Replies.map(reply => ({
         ...reply
       }))
 
-      res.render('tweet', { tweetReplies, tweet: tweetJson, popularUser })
+      res.render('tweet', {status: (200), tweetReplies, tweet: tweetJson, popularUser })
     } catch (err) {
-      console.warn(err)
+      console.log(err)
+      res.status(302);
+      console.log('getTweet err')
+      req.flash('error_messages', '讀取貼文失敗')
+      return res.redirect('back')
     }
   },
   postReplies: async (req, res) => {
@@ -88,37 +103,50 @@ const tweetController = {
       }
       const reply = await Reply.create({
         comment,
-        UserId: req.user.id,
+        UserId: helpers.getUser(req).id,
         TweetId: req.params.tweetId
       })
+      res.status(200)
       return res.redirect(`/tweets/${req.params.tweetId}/replies`)
     } catch (err) {
-      console.warn(err)
+      console.log(err)
+      res.status(302)
+      console.log('PostReplies err')
+      req.flash('error_messages', '留言失敗')
+      return res.redirect('back')
     }
   },
   addLike: async (req, res) => {
     try {
       await Like.create({
-        UserId: req.user.id,
+        UserId: helpers.getUser(req).id,
         TweetId: req.params.tweetId
       })
       return res.redirect('back')
     } catch (err) {
-      console.warn(err)
+      console.log(err)
+      res.status(302);
+      console.log('addLike err')
+      req.flash('error_messages', '點擊失敗')
+      return res.redirect('back')
     }
   },
   removeLike: async (req, res) => {
     try {
-      console.log('in removeLike controller')
-      await Like.destroy({
-        where: {
-          UserId: req.user.id,
-          TweetId: req.params.tweetId
-        }
-      })
+      const UserId = helpers.getUser(req).id
+      const TweetId = Number(req.params.tweetId)
+      console.log(UserId, TweetId)
+
+      await Like.destroy({ where: { UserId, TweetId } })
+
       return res.redirect('back')
+
     } catch (err) {
-      console.warn(err)
+      console.log(err)
+      res.status(302);
+      console.log('removeLike err')
+      req.flash('error_messages', '點擊失敗')
+      return res.redirect('back')
     }
   }
 }
